Modify the ncurses UI implementation to a FSM
Modified the ncurses code to a FSM, this should have been done earlier.
States are implemented as functions which takes 2 arguments 1st is
a pointer to the ncurses window and 2nd is an optional void** to
additional data. State functions return function pointer to the next
state. Currently there is only one state.